+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Business Support module / Objet TeamTask : Menu entries for customers.Added new role(role.customer) in data-init.
 - Employee:Company(Employee) Phonebook 
 - User Form : Provide step wise view.
+- Base : Addition DMS Import Feature.
 
 ## Improvements
 - TimesheetLine: Adding M2O field TeamTask and integer field timeToCharge.